Conflicts:
	Changelog.md
	Roadmap.md
	erogenousbeef/bigreactors/client/gui/BeefGuiBase.java
	erogenousbeef/bigreactors/common/BRConfig.java
	erogenousbeef/bigreactors/common/multiblock/MultiblockReactor.java
	erogenousbeef/bigreactors/gui/controls/BeefGuiProgressBarVertical.java
	mcmod.info

+++ erogenousbeef/bigreactors/api/HeatPulse.java
@@ -1,11 +1,11 @@
 package erogenousbeef.bigreactors.api;
 
 public class HeatPulse {
-	public double powerProduced;
-	public double heatChange;
+	public float powerProduced;
+	public float heatChange;
 	
 	public HeatPulse() {
-		powerProduced = 0.0;
-		heatChange = 0.0;
+		powerProduced = 0.0f;
+		heatChange = 0.0f;
 	}
 }
